New per app state definitions
